refactor(preferences): read the context from its own module in the remaining pages
